Bundle scripts and update HTML
