+++ dones.html
@@ -77,7 +77,8 @@
   </div>
 
   <script src="js/bundle-utils-1.js"></script>
-  <script type="module" src="js/dones.js"></script>
+  <script src="js/bundle-dones.js"></script>
+  <script src="js/dones.js"></script>
   <script src="js/bundle-auth-nav.js"></script>
   <script>
     // Inicializar autenticación cuando el DOM esté listo
@@ -89,6 +90,5 @@
   </script>
   
   <script src="js/feedback-modal.js"></script>
-  <script src="js/dones-tabs.js"></script>
 </body>
 </html>
