+++ forja-mistica.html
@@ -240,7 +240,7 @@
       toggle(activeBtn.dataset.tab);
     });
   </script>
-  <script type="module" src="js/forja-mistica.js"></script>
+  <script src="js/bundle-forja-mistica.js"></script>
   <script src="js/bundle-auth-nav.js"></script>
   <script>
     // Inicializar autenticación cuando el DOM esté listo
