+++ js/bundle-forja-mistica.js
@@ -1,4 +1,52 @@
-import { fetchItemPrices, fetchIconsFor, iconCache } from './fractales-utils.js';
+// Bundled forja mistica scripts
+// Utilidades compartidas para fractales y forja mística
+const iconCache = {};
+const rarityCache = {};
+
+async function fetchIconsFor(ids = []) {
+  if (!ids.length) return;
+  try {
+    const res = await fetch(`https://api.guildwars2.com/v2/items?ids=${ids.join(',')}&lang=es`);
+    const data = await res.json();
+    data.forEach(item => {
+      if (item && item.id) {
+        iconCache[item.id] = item.icon;
+        rarityCache[item.id] = item.rarity;
+      }
+    });
+  } catch {}
+}
+
+async function fetchItemPrices(ids = []) {
+  if (!ids || ids.length === 0) return {};
+  const url = `https://api.datawars2.ie/gw2/v1/items/csv?fields=id,buy_price,sell_price&ids=${ids.join(',')}`;
+  try {
+    const csv = await fetch(url).then(r => r.text());
+    const [header, ...rows] = csv.trim().split('\n');
+    const headers = header.split(',');
+    const idIdx = headers.indexOf('id');
+    const buyIdx = headers.indexOf('buy_price');
+    const sellIdx = headers.indexOf('sell_price');
+    const result = {};
+    rows.forEach(row => {
+      const cols = row.split(',');
+      const id = parseInt(cols[idIdx], 10);
+      if (!isNaN(id)) {
+        result[id] = {
+          buy_price: parseInt(cols[buyIdx], 10) || 0,
+          sell_price: parseInt(cols[sellIdx], 10) || 0
+        };
+      }
+    });
+    return result;
+  } catch (e) {
+    return {};
+  }
+}
+
+if (typeof window !== 'undefined') {
+  window.FractalesUtils = { fetchIconsFor, fetchItemPrices, iconCache, rarityCache };
+}
 
 
 function addIconToCell(cell, icon) {
@@ -11,7 +59,7 @@ function addIconToCell(cell, icon) {
   div.prepend(img);
 }
 
-export const MATERIAL_IDS = {
+const MATERIAL_IDS = {
   t6: {
     sangre: 24295,
     hueso: 24358,
@@ -34,7 +82,7 @@ export const MATERIAL_IDS = {
   piedra: 20796
 };
 
-export const LODESTONE_IDS = {
+const LODESTONE_IDS = {
   cores: {
     glacial: 24319,
     cristal: 24329,
@@ -58,7 +106,7 @@ export const LODESTONE_IDS = {
   cristal: 20799
 };
 
-export async function renderTablaForja() {
+async function renderTablaForja() {
   const keys = Object.keys(MATERIAL_IDS.t5);
   const ids = [
     ...keys.map(k => MATERIAL_IDS.t5[k]),
@@ -99,7 +147,7 @@ export async function renderTablaForja() {
   });
 }
 
-export async function renderTablaLodestones() {
+async function renderTablaLodestones() {
   const coreKeys = Object.keys(LODESTONE_IDS.cores);
   const ids = [
     ...coreKeys.map(k => LODESTONE_IDS.cores[k]),
