--- js/data/basicMaterials1gen.js
@@ -1,191 +1,0 @@
-// Mapeo de materiales básicos, dones y precursores para legendarios
-// Fuente: https://wiki-es.guildwars2.com/wiki/Crep%C3%BAsculo
-
-export const BASIC_MATERIALS = {
-// === MATERIALES T6 ===
-24295: { name: 'Vial de sangre poderosa', type: 'crafting_material' },
-24283: { name: 'Vesícula de veneno poderoso', type: 'crafting_material' },
-24300: { name: 'Tótem elaborado', type: 'crafting_material' },
-24277: { name: 'Montón de polvo cristalino', type: 'crafting_material' },
-24351: { name: 'Colmillo feroz', type: 'crafting_material' },
-24289: { name: 'Escama blindada', type: 'crafting_material' },
-24357: { name: 'Garra despiadada', type: 'crafting_material' },
-24358: { name: 'Hueso antiguo', type: 'crafting_material' },
-19721: { name: 'Pegote de ectoplasma', type: 'crafting_material' }, // Añadido para asegurar el precio
-
-// === LEGENDARIOS ===
-30684: { name: 'Colmilloescarcha', type: 'legendary' },
-30687: { name: 'Incineradora', type: 'legendary' },
-30692: { name: 'El Festín', type: 'legendary' },
-30695: { name: 'Meteorológico', type: 'legendary' },
-30701: { name: 'Kraitkin', type: 'legendary' },
-30691: { name: 'Kamhoali\'i Kotaki', type: 'legendary' },
-30697: { name: 'Frenesí', type: 'legendary' },
-30698: { name: 'El Bifrost', type: 'legendary' },
-30686: { name: 'El Soñador', type: 'legendary' },
-30694: { name: 'El Depredador', type: 'legendary' },
-30685: { name: 'Kudzu', type: 'legendary' },
-30690: { name: 'El Juggernaut', type: 'legendary' },
-30689: { name: 'Eternidad', type: 'legendary' },
-30703: { name: 'Amanecer', type: 'legendary' },
-30702: { name: 'Aullador', type: 'legendary' },
-30700: { name: 'Rodgort', type: 'legendary' },
-30696: { name: 'Las Profecías del Buscador de la Llama', type: 'legendary' },
-30688: { name: 'El Juglar', type: 'legendary' }, 
-30699: { name: 'Haz', type: 'legendary' },
-30704: { name: 'Crepúsculo', type: 'legendary' },
-
-
-// === DONES ===
-19672: { name: 'Don del poder', type: 'account_bound' },
-19673: { name: 'Don de la magia', type: 'account_bound' },
-19677: { name: 'Don de la exploración', type: 'account_bound' },
-19678: { name: 'Don de la batalla', type: 'account_bound' },
-19664: { name: 'Don de Ascalon', type: 'account_bound' },
-19674: { name: 'Don del dominio', type: 'account_bound' },
-19648: { name: 'Don del Crepúsculo', type: 'crafting_material' },
-19621: { name: 'Don del metal', type: 'crafting_material' },
-19631: { name: 'Don de la oscuridad', type: 'crafting_material' },
-19626: { name: 'Don de la suerte', type: 'crafting_material' },
-19665: { name: 'Don del noble', type: 'crafting_material' },
-19635: { name: 'Don del entretenimiento', type: 'crafting_material' },
-19650: { name: 'Don del Festín', type: 'crafting_material' },
-19645: { name: 'Don de Incineradora', type: 'crafting_material' },
-19624: { name: 'Don del hielo', type: 'crafting_material' },
-19670: { name: 'Don del Santuario', type: 'crafting_material' },
-19657: { name: 'Don de Kamohoali\'i Kotaki', type: 'crafting_material' },
-19669: { name: 'Don de Zhaitan', type: 'crafting_material' },
-19659: { name: 'Don del Frenesí', type: 'crafting_material' },
-19643: { name: 'Don del agua', type: 'crafting_material' },
-19638: { name: 'Don del color', type: 'crafting_material' },
-19654: { name: 'Don del Bifrost', type: 'crafting_material' },
-19660: { name: 'Don del Soñador', type: 'crafting_material' },
-19628: { name: 'Estatua de unicornio', type: 'crafting_material' },
-19667: { name: 'Don de las espinas', type: 'crafting_material' },
-19661: { name: 'Don del Depredador', type: 'crafting_material' },
-19671: { name: 'Don del conocimiento', type: 'crafting_material' },
-19636: { name: 'Don del sigilo', type: 'crafting_material' },
-19647: { name: 'Don del Amanecer', type: 'crafting_material' },
-19632: { name: 'Don de la luz', type: 'crafting_material' },
-19662: { name: 'Don de Aullador', type: 'crafting_material' },
-19627: { name: 'Don de la naturaleza', type: 'crafting_material' },
-19644: { name: 'Don de Kudzu', type: 'crafting_material' },
-19649: { name: 'Don del Juggernaut', type: 'crafting_material' },
-19625: { name: 'Don de Colmilloescarcha', type: 'crafting_material' },
-19651: { name: 'Don de Gracia', type: 'crafting_material' },
-19622: { name: 'Don de la madera', type: 'crafting_material' },
-19666: { name: 'Don del herrador', type: 'crafting_material' },
-19658: { name: 'Don de Kraitkin', type: 'crafting_material' },
-19623: { name: 'Don de la energía', type: 'crafting_material' },
-19656: { name: 'Don de Rodgort', type: 'crafting_material' },
-19653: { name: 'Don de Las Profecías del Buscador de la Llama', type: 'crafting_material' },
-19629: { name: 'Don de la historia', type: 'crafting_material' },
-19646: { name: 'Don del Juglar', type: 'crafting_material' },
-19630: { name: 'Don de la música', type: 'crafting_material' },
-19655: { name: 'Don de Haz', type: 'crafting_material' },
-19639: { name: 'Don del relámpago', type: 'crafting_material' },
-19652: { name: 'Don del Meteorológico', type: 'crafting_material' },
-19637: { name: 'Don del clima', type: 'crafting_material' },
-19641: { name: 'Estatua de tiburón', type: 'crafting_material' },
-
-  // === LINGOTES ===
-  19688: { name: 'Lingote de acero', type: 'crafting_material' },
-  19685: { name: 'Lingote de oricalco', type: 'crafting_material' },
-  19684: { name: 'Lingote de mithril', type: 'crafting_material' },
-  19683: { name: 'Lingote de hierro', type: 'crafting_material' },
-  19681: { name: 'Lingote de aceroscuro', type: 'crafting_material' },
-  19686: { name: 'Lingote de platino', type: 'crafting_material' },
-
-// === SELLOS ===
-24592: { name: 'Sello superior de vitalidad', type: 'upgrade_component' },
-24632: { name: 'Sello superior de veneno', type: 'upgrade_component' },
-24612: { name: 'Sello superior de agonía', type: 'upgrade_component' },
-24561: { name: 'Sello superior de rabia', type: 'upgrade_component' },
-24572: { name: 'Sello superior de invalidación', type: 'upgrade_component' },
-24571: { name: 'Sello superior de pureza', type: 'upgrade_component' },
-24615: { name: 'Sello superior de fuerza', type: 'upgrade_component' },
-24618: { name: 'Sello superior de precisión', type: 'upgrade_component' },
-24601: { name: 'Sello superior de batalla', type: 'upgrade_component' },
-24554: { name: 'Sello superior de aire', type: 'upgrade_component' },
-24584: { name: 'Sello superior de benevolencia', type: 'upgrade_component' },
-24562: { name: 'Sello superior de fortaleza', type: 'upgrade_component' },
-24865: { name: 'Sello superior de celeridad', type: 'upgrade_component' },
-24651: { name: 'Sello superior de sangre', id: 24570, type: 'upgrade_component' },
-24555: { name: 'Sello superior de hielo', type: 'upgrade_component' },
-24548: { name: 'Sello superior de fuego', type: 'upgrade_component' },
-24607: { name: 'Sello superior de energía', type: 'upgrade_component' },
-
-  // === PIEDRAS IMÁN ===
-  24310: { name: 'Piedra imán de ónice', type: 'crafting_material' },
-  24320: { name: 'Piedra imán glacial', type: 'crafting_material' },
-  24340: { name: 'Piedra imán corrupta', type: 'crafting_material' },
-  24315: { name: 'Piedra imán fundida', type: 'crafting_material' },
-  24325: { name: 'Piedra imán de destructor', type: 'crafting_material' },
-  24305: { name: 'Piedra imán cargada', type: 'crafting_material' },
-  
-
-  // === TABLAS DE MADERA ===
-  19709: { name: 'Tabla de madera ancestral', type: 'crafting_material' },
-  19711: { name: 'Tabla de madera sólida', type: 'crafting_material' },
-  19712: { name: 'Tabla de madera antigua', type: 'crafting_material' },
-  19714: { name: 'Tabla de madera curtida', type: 'crafting_material' },
-  19728: { name: 'Tabla de madera de arista', type: 'crafting_material' },
-
-  // === ORBES ===
-  24522: { name: 'Orbe de ópalo', type: 'crafting_material' },
-  24512: { name: 'Orbe de crisocola', type: 'crafting_material' },
-
-  // === POLVOS ===
-  24274: { name: 'Montón de polvo radiante', type: 'crafting_material' },
-  24276: { name: 'Montón de polvo incandescente', type: 'crafting_material' },
-  24275: { name: 'Montón de polvo luminoso', type: 'crafting_material' },
-  24277: { name: 'Montón de polvo cristalino', type: 'crafting_material' },
-
-  // === PRECURSORES ===
-  29183: { name: 'Veneno', type: 'weapon' },
-  29171: { name: 'Carcharias', type: 'weapon' },
-  29180: { name: 'La Leyenda', type: 'weapon' },
-  29178: { name: 'El Amante', type: 'weapon' },
-  29175: { name: 'El Cazador', type: 'weapon' },
-  29170: { name: 'El Coloso', type: 'weapon' },
-  29184: { name: 'Aullido', type: 'weapon' },
-  29177: { name: 'El Elegido', type: 'weapon' },
-  29168: { name: 'El Bardo', type: 'weapon' },
-  29181: { name: 'Zas', type: 'weapon' },
-  29176: { name: 'Tormenta', type: 'weapon' },
-  29172: { name: 'Hoja de Kudzu', type: 'weapon' },
-  29167: { name: 'Chispa', type: 'weapon' },
-  29166: { name: 'Diente de Colmilloescarcha', type: 'weapon' },
-  29173: { name: 'El Energizador', type: 'weapon' },
-  29174: { name: 'Pistola del Caos', type: 'weapon' },
-  29169: { name: 'Alba', type: 'weapon' },
-  29182: { name: 'Llama de Rodgort', type: 'weapon' },
-  29185: { name: 'Anochecer', type: 'weapon' },
-
-  
-  // === ESTATUAS ===
-  19642: { name: 'Estatua de anguila', type: 'crafting_material' },
-  19640: { name: 'Estatua de lobo', type: 'crafting_material' },
-  19628: { name: 'Estatua de unicornio', type: 'crafting_material' },
-
-
-  // === OTROS ===
-  12545: { name: 'Trufa orriana', type: 'crafting_material' },
-  12128: { name: 'Baya omnom', type: 'crafting_material' },
-  19732: { name: 'Trozos de cuero endurecido', type: 'crafting_material' },
-  19633: { name: 'Vial de azogue', type: 'crafting_material' },
-  19634: { name: 'Vial de llama líquida', type: 'crafting_material' },
-  20323: { name: 'Tinte sin identificar', type: 'crafting_material' },
-  19721: { name: 'Pegote de ectoplasma', type: 'crafting_material' },
-  19925: { name: 'Esquirla de obsidiana', type: 'crafting_material' },
-  20797: { name: 'Esquirla de hematites', type: 'crafting_material' },
-  20796: { name: 'Piedra filosofal', type: 'crafting_material' },
-  19976: { name: 'Moneda mística', type: 'crafting_material' },
-  19676: { name: 'Piedra rúnica helada', type: 'crafting_material' },
-  12544: { name: 'Chile fantasma', type: 'crafting_material' },
-  19746: { name: 'Haz de gasa', type: 'crafting_material' },
-  24502: { name: 'Doblón de plata', type: 'crafting_material' }, 
-  19675: { name: 'Trébol místico', type: 'crafting_material' },
-  19737: { name: 'Retal de cuero curado endurecido', type: 'crafting_material' },
-  20000: { name: 'Caja de diversión', type: 'crafting_material' },
-};
