--- js/data/basicMaterials3gen.js
@@ -1,204 +1,0 @@
-// Materiales básicos para legendarias de 3ra generación
-// Basado en la legendaria Desgarro de Aurene
-
-export const BASIC_MATERIALS_3GEN = {
-
-// === DON DE JADE ===
-// === DON DE JADE ===
-96033: { name: 'Don de dominio de jade', type: 'crafting_material' },
-
-  // === MATERIALES DE JADE ===
-  97433: { name: 'Don del Imperio del Dragón', type: 'crafting_material' },
-  96722: { name: 'Piedra rúnica de jade', type: 'crafting_material' },
-  97102: { name: 'Pedazo de jade puro', type: 'crafting_material' },
-  96347: { name: 'Pedazo de ámbar gris antiguo', type: 'crafting_material' },
-  97829: { name: 'Bendición de la Emperatriz de Jade', type: 'account_bound' },
-
-  // === OTROS DE JADE===
-  20797: { name: 'Esquirla de hematites', type: 'crafting_material' },
-  96978: { name: 'Piedra de invocación vetusta', type: 'crafting_material' },
-
-  // === MATERIALES DON DE CANTHA ===
-  97096: { name: 'Don de Cantha', type: 'account_bound' },
-  96993: { name: 'Don de la Provincia de Seitung', type: 'account_bound' },
-  95621: { name: 'Don de la ciudad de Nueva Kaineng', type: 'account_bound' },
-  97232: { name: 'Don del Bosque Echovald', type: 'account_bound' },
-  96083: { name: 'Don de Muerte del Dragón', type: 'account_bound' },
-
-  // === DON DE DRACÓNICO ===
-  // === DON DE DRACÓNICO ===
-  96137: { name: 'Tributo dracónico', type: 'crafting_material' },
-
-  // === MATERIALES TRÉBOL ===
-  19675: { name: 'Trébol místico', type: 'crafting_material' },
-  19976: { name: 'Moneda mística', type: 'currency' },
-  19721: { name: 'Pegote de ectoplasma', type: 'crafting_material' },
-  19925: { name: 'Esquirla de obsidiana', type: 'crafting_material' },
-  20799: { name: 'Cristal místico', type: 'crafting_material' },
-  20796: { name: 'Piedra filosofal', type: 'crafting_material' },
-
-  // === DONES CONDENSADOS ===
-  // Poder
-  70867: { name: 'Don de poder condensado', type: 'account_bound' },
-  70801: { name: 'Don de garras', type: 'account_bound' },
-  75299: { name: 'Don de escamas', type: 'account_bound' },
-  71123: { name: 'Don de huesos', type: 'account_bound' },
-  75744: { name: 'Don de colmillos', type: 'account_bound' },
-  // Magia
-  76530: { name: 'Don de magia condensada', type: 'account_bound' },
-  71655: { name: 'Don de sangre', type: 'account_bound' },
-  71787: { name: 'Don de veneno', type: 'account_bound' },
-  73236: { name: 'Don de tótems', type: 'account_bound' },
-  73196: { name: 'Don de polvo', type: 'account_bound' },
-
-  // === OTROS DE DRACÓNICO===
-  92687: { name: 'Piedra imán dracónica amalgamada', type: 'crafting_material' },
-
-  // === DON DE AURENE CAMBIANTE ===
-  // === DON DE AURENE CAMBIANTE ===
-
-  71581: { name: 'Memoria de batalla', type: 'crafting_material' },
-  70820: { name: 'Esquirla de gloria', type: 'crafting_material' },
-  
-  // === MATERIALES DON DE INVESTIGACIÓN ===
-  97655: { name: 'Don de investigación', type: 'crafting_material' },
-  46747: { name: 'Reactivo termocatalítico', type: 'crafting_material' },
-  95813: { name: 'Reactivo hidrocatalítico', type: 'crafting_material' },
-  45178: { name: 'Esencia de la suerte exótica', type: 'currency' },
-
-  // === MATERIALES DON DE LA NIEBLA ===
-  76427: { name: 'Don de la niebla', type: 'account_bound' },
-  70528: { name: 'Don de gloria', type: 'account_bound' },
-  19678: { name: 'Don de la batalla', type: 'account_bound' },
-  71008: { name: 'Don de guerra', type: 'account_bound' },
-  73137: { name: 'Cubo de energía oscura estabilizada', type: 'crafting_material' },
-  71994: { name: 'Bola de energía oscura', type: 'crafting_material' },
-  73248: { name: 'Matriz estabilizadora', type: 'crafting_material' },
-
-  // === MATERIALES DE POEMA FIJO ===
-  96151: { name: 'Relato de aventura', type: 'crafting_material' },
-  97790: { name: 'Insignia de farolero', type: 'crafting_material' },
-  71148: { name: 'Hoja de papel supremo', type: 'crafting_material' },
-
-  // === POEMAS SEGUN ARMA ===
-  97160: { name: 'Poema sobre hachas', type: 'crafting_material' }, // Hacha
-  96187: { name: 'Poema sobre dagas', type: 'crafting_material' }, // Daga
-  96035: { name: 'Poema sobre mazas', type: 'crafting_material' }, // Maza
-  95809: { name: 'Poema sobre pistolas', type: 'crafting_material' }, // Pistola
-  96173: { name: 'Poema sobre cetros', type: 'crafting_material' }, // Cetro
-  97335: { name: 'Poema sobre espadas', type: 'crafting_material' }, // Espada
-  96951: { name: 'Poema sobre focos', type: 'crafting_material' }, // Foco
-  95740: { name: 'Poema sobre escudos', type: 'crafting_material' }, // Escudo
-  97257: { name: 'Poema sobre antorchas', type: 'crafting_material' }, // Antorcha
-  96341: { name: 'Poema sobre cuernos de guerra', type: 'crafting_material' }, // Cuerno de guerra
-  96036: { name: 'Poema sobre mandobles', type: 'crafting_material' }, // Mandoble
-  97082: { name: 'Poema sobre martillos', type: 'crafting_material' }, // Martillo
-  97800: { name: 'Poema sobre arcos largos', type: 'crafting_material' }, // Arco largo
-  97201: { name: 'Poema sobre rifles', type: 'crafting_material' }, // Rifle
-  96849: { name: 'Poema sobre arcos cortos', type: 'crafting_material' }, // Arco corto
-  95962: { name: 'Poema sobre bastones', type: 'crafting_material' }, // Báculo
-  
-  // === OTROS AURENE FIJOS===
-  79418: { name: 'Piedra rúnica mística', type: 'crafting_material' },
-
-// === MATERIALES ORDENADOS POR TIPO DE ARMA ===
-45845: { name: 'Filo para hacha de acero de Deldrimor', type: 'crafting_material' },          // Hacha
-45846: { name: 'Filo para daga de acero de Deldrimor', type: 'crafting_material' },           // Daga
-45852: { name: 'Cabeza para maza de acero de Deldrimor', type: 'crafting_material' },         // Maza
-45833: { name: 'Cañón para pistola de acero de Deldrimor', type: 'crafting_material' },       // Pistola
-45885: { name: 'Núcleo para cetro de madera espiritual', type: 'crafting_material' },         // Cetro
-45848: { name: 'Filo para espada de acero de Deldrimor', type: 'crafting_material' },         // Espada
-45884: { name: 'Núcleo para foco de madera espiritual', type: 'crafting_material' },          // Foco
-45858: { name: 'Umbo para escudo de acero de Deldrimor', type: 'crafting_material' },         // Escudo
-45838: { name: 'Cabezal para antorcha de acero de Deldrimor', type: 'crafting_material' },    // Antorcha
-45839: { name: 'Cuerno de acero de Deldrimor', type: 'crafting_material' },                   // Cuerno de guerra
-45847: { name: 'Filo para mandoble de acero de Deldrimor', type: 'crafting_material' },       // Mandoble
-45851: { name: 'Cabeza para martillo de acero de Deldrimor', type: 'crafting_material' },     // Martillo
-45841: { name: 'Duela para arco largo de madera espiritual', type: 'crafting_material' },     // Arco largo
-45834: { name: 'Cañón para rifle de acero de Deldrimor', type: 'crafting_material' },         // Rifle
-45842: { name: 'Duela para arco corto de madera espiritual', type: 'crafting_material' },     // Arco corto
-45887: { name: 'Cabezal para báculo de madera espiritual', type: 'crafting_material' },        // Báculo
-
-
-// === PRECURSORES ORDENADOS ===
-97449: { name: 'Desgarro Dracónico (precursora)', type: 'weapon' }, // Hacha
-95967: { name: 'Garra Dracónica (precursora)', type: 'weapon' }, // Daga
-96827: { name: 'Cola Dracónica (precursora)', type: 'weapon' }, // Maza
-96915: { name: 'Razonamiento Dracónico (precursora)', type: 'weapon' }, // Pistola
-96193: { name: 'Sabiduría Dracónica (precursora)', type: 'weapon' }, // Cetro
-95994: { name: 'Colmillo Dracónico (precursora)', type: 'weapon' }, // Espada
-96303: { name: 'Mirada Dracónica (precursora)', type: 'weapon' }, // Foco
-97691: { name: 'Escama Dracónica (precursora)', type: 'weapon' }, // Escudo
-96925: { name: 'Aliento Dracónico (precursora)', type: 'weapon' }, // Antorcha
-97513: { name: 'Voz Dracónica (precursora)', type: 'weapon' }, // Cuerno de guerra
-96357: { name: 'Mordisco Dracónico (precursora)', type: 'weapon' }, // Mandoble
-95920: { name: 'Peso Dracónico (precursora)', type: 'weapon' }, // Martillo
-95834: { name: 'Vuelo Dracónico (precursora)', type: 'weapon' }, // Arco largo
-97267: { name: 'Persuasión Dracónica (precursora)', type: 'weapon' }, // Rifle
-96330: { name: 'Ala Dracónica (precursora)', type: 'weapon' }, // Arco corto
-95814: { name: 'Reflexión Dracónica (precursora)', type: 'weapon' }, // Báculo
-
-
-// === LEGENDARIAS ORDENADAS ===
-96937: { name: 'Desgarro de Aurene', type: 'legendary' }, // Hacha
-96203: { name: 'Garra de Aurene', type: 'legendary' }, // Daga
-95612: { name: 'Cola de Aurene', type: 'legendary' }, // Maza
-95808: { name: 'Razonamiento de Aurene', type: 'legendary' }, // Pistola
-96221: { name: 'Sabiduría de Aurene', type: 'legendary' }, // Cetro
-95675: { name: 'Colmillo de Aurene', type: 'legendary' }, // Espada
-97165: { name: 'Mirada de Aurene', type: 'legendary' }, // Foco
-96028: { name: 'Escama de Aurene', type: 'legendary' }, // Escudo
-97099: { name: 'Aliento de Aurene', type: 'legendary' }, // Antorcha
-97783: { name: 'Voz de Aurene', type: 'legendary' }, // Cuerno de guerra
-96356: { name: 'Mordisco de Aurene', type: 'legendary' }, // Mandoble
-95684: { name: 'Peso de Aurene', type: 'legendary' }, // Martillo
-97590: { name: 'Vuelo de Aurene', type: 'legendary' }, // Arco largo
-97377: { name: 'Persuasión de Aurene', type: 'legendary' }, // Rifle
-97077: { name: 'Ala de Aurene', type: 'legendary' }, // Arco corto
-96652: { name: 'Reflexión de Aurene', type: 'legendary' }, // Báculo
-
-
-  
-  // ====================
-  // === MATERIALES T6 ===
-  24295: { name: 'Vial de sangre poderosa', type: 'crafting_material' },
-  24283: { name: 'Vesícula de veneno poderoso', type: 'crafting_material' },
-  24300: { name: 'Tótem elaborado', type: 'crafting_material' },
-  24277: { name: 'Montón de polvo cristalino', type: 'crafting_material' },
-  24357: { name: 'Colmillo feroz', type: 'crafting_material' },
-  24289: { name: 'Escama blindada', type: 'crafting_material' },
-  24351: { name: 'Garra despiadada', type: 'crafting_material' },
-  24358: { name: 'Hueso antiguo', type: 'crafting_material' },
-
-  // === MATERIALES T5 ===
-  24294: { name: 'Vial de sangre potente', type: 'crafting_material' },
-  24282: { name: 'Vesícula de veneno potente', type: 'crafting_material' },
-  24299: { name: 'Tótem intrincado', type: 'crafting_material' },
-  24276: { name: 'Montón de polvo incandescente', type: 'crafting_material' },
-  24356: { name: 'Colmillo grande', type: 'crafting_material' },
-  24288: { name: 'Escama grande', type: 'crafting_material' },
-  24350: { name: 'Garra grande', type: 'crafting_material' },
-  24341: { name: 'Hueso grande', type: 'crafting_material' },
-
-  // === MATERIALES T4 ===
-  24293: { name: 'Vial de sangre espesa', type: 'crafting_material' },
-  24281: { name: 'Vesícula de veneno llena', type: 'crafting_material' },
-  24298: { name: 'Tótem grabado', type: 'crafting_material' },
-  24275: { name: 'Montón de polvo luminoso', type: 'crafting_material' },
-  24355: { name: 'Colmillo afilado', type: 'crafting_material' },
-  24287: { name: 'Escama suave', type: 'crafting_material' },
-  24349: { name: 'Garra afilada', type: 'crafting_material' },
-  24345: { name: 'Hueso pesado', type: 'crafting_material' },
-
-  // === MATERIALES T3 ===
-  24292: { name: 'Vial de sangre', type: 'crafting_material' },
-  24280: { name: 'Vesícula de veneno', type: 'crafting_material' },
-  24297: { name: 'Tótem', type: 'crafting_material' },
-  24274: { name: 'Montón de polvo radiante', type: 'crafting_material' },
-  24354: { name: 'Colmillo', type: 'crafting_material' },
-  24286: { name: 'Escama', type: 'crafting_material' },
-  24348: { name: 'Garra', type: 'crafting_material' },
-  24344: { name: 'Hueso', type: 'crafting_material' }
-
-
-};
