--- js/legendaryCraftingBase.js
@@ -1,1 +1,0 @@
-export { LegendaryCraftingBase } from './legendary-utils.js';
