+++ leg-craft.html
@@ -134,6 +134,7 @@
 
   <script src="js/bundle-utils-1.js"></script>
   <script src="js/bundle-auth-nav.js"></script>
+  <script src="js/bundle-legendary.js"></script>
   <script>
     // Inicializar autenticación cuando el DOM esté listo
     document.addEventListener('DOMContentLoaded', function() {
@@ -143,7 +144,7 @@
     });
   </script>
   
-  <script type="module" src="js/leg-craft-tabs.js"></script>
+  <script src="js/leg-craft-tabs.js"></script>
   
   <script src="js/feedback-modal.js"></script>
 </body>
